Fix geek-rh.eclass, geek-sources.eclass, sys-kernel/geek-sources/ChangeLog Add USE="rh" to sys-kernel/geek-sources-3.10.32

+++ eclass/geek-rh.eclass
@@ -40,6 +40,7 @@ EXPORT_FUNCTIONS src_unpack src_prepare pkg_postinst
 geek-rh_init_variables() {
 	debug-print-function ${FUNCNAME} "$@"
 
+if [[ ${KMV} = "2.6" ]]; then
 	: ${RH_VER:=${RH_VER:-"19"}} # Patchset version based on kernel-2.6.32-19.el6.src.rpm
 	: ${RH_NAME:=${RH_NAME:-kernel-${VERSION}.${PATCHLEVEL}.${SUBLEVEL}-${RH_VER/KMV/$KMV}.el6}}
 	SRC_URI_PREFIX="http://ftp.redhat.com/pub/redhat/linux/enterprise"
@@ -47,7 +48,13 @@ geek-rh_init_variables() {
 	${SRC_URI_PREFIX}/6ComputeNode/en/os/SRPMS/${RH_NAME}.src.rpm
 	${SRC_URI_PREFIX}/6Server/en/os/SRPMS/${RH_NAME}.src.rpm
 	${SRC_URI_PREFIX}/6Workstation/en/os/SRPMS/${RH_NAME}.src.rpm"}} # Patchset sources url
-
+elif [[ ${KMV} = "3.10" ]]; then
+	: ${RH_VER:=${RH_VER:-"54.0.1"}} # Patchset version based on kernel-3.10.0-54.0.1.el7.src.rpm
+	: ${RH_NAME:=${RH_NAME:-kernel-3.10.0-${RH_VER/KMV/$KMV}.el7}}
+	SRC_URI_PREFIX="http://ftp.redhat.com/redhat/rhel/beta"
+	: ${RH_SRC:=${RH_SRC:-"${SRC_URI_PREFIX}/7/source/SRPMS/${RH_NAME}.src.rpm"}} # Patchset sources url
+	SKIP_UPDATE="1"
+fi
 	: ${RH_URL:=${RH_URL:-"http://www.redhat.com"}} # Patchset url
 	: ${RH_INF:="${YELLOW}Red Hat Enterprise Linux kernel patches version ${GREEN}${RH_VER}${YELLOW} from ${GREEN}${RH_URL}${NORMAL}"}
 
@@ -77,26 +84,39 @@ geek-rh_src_unpack() {
 	shift
 	test -d "${CWD}" >/dev/null 2>&1 && cd "${CWD}" || mkdir -p "${CWD}"; cd "${CWD}"
 
-	rpm_unpack "${RH_NAME}.src.rpm" || die
+	if [[ ${KMV} = "2.6" ]]; then
+		rpm_unpack "${RH_NAME}.src.rpm" || die
 
-	if [ -e "${WORKDIR}/${RH_NAME}.tar.bz2" ]; then
-		tar -xpf "${WORKDIR}/${RH_NAME}.tar.bz2" || die
-		mv "${RH_NAME}" "${S}" || die
-		rm -f "${WORKDIR}/${RH_NAME}.tar.bz2" || die
+		if [ -e "${WORKDIR}/${RH_NAME}.tar.bz2" ]; then
+			tar -xpf "${WORKDIR}/${RH_NAME}.tar.bz2" || die
+			mv "${RH_NAME}" "${S}" || die
+			rm -f "${WORKDIR}/${RH_NAME}.tar.bz2" || die
+		fi
+
+		# Delete crap patches from rh patchset
+		local rh_crap_patch="linux-2.6.32.tar.bz2 redhat-Include-FIPS-required-checksum-of-the-kernel-image.patch redhat-Silence-tagging-messages-by-rh-release.patch redhat-Disabling-debug-options-for-beta.patch redhat-updating-lastcommit-for-2-6-31-50.patch redhat-fix-BZ-and-CVE-info-printing-on-changelog-when-HIDE_REDHAT-is-enabled.patch redhat-updating-lastcommit-for-2-6-31-51.patch redhat-Fix-version-passed-to-update_changelog-sh.patch redhat-updating-lastcommit-for-2-6-32-0-52.patch redhat-fix-STAMP-version-on-rh-release-commit-phase.patch redhat-enable-debug-builds-also-on-s390x-and-ppc64.patch redhat-updating-lastcommit-for-2-6-32-0-53.patch redhat-fixing-wrong-bug-number-536759-536769.patch redhat-adding-top-makefile-to-enable-rh-targets.patch redhat-updating-lastcommit-for-2-6-32-0-54.patch redhat-updating-lastcommit-for-2-6-32-1.patch redhat-update-build-targets-in-Makefile.patch redhat-include-missing-System-map-file-for-debug-only-builds.patch redhat-updating-lastcommit-for-2-6-32-2.patch redhat-force-to-run-rh-key-target-when-compiling-the-kernel-locally-without-RPM.patch redhat-fixing-lastcommit-contents-for-2-6-32-2-el6.patch redhat-updating-lastcommit-for-2-6-32-3.patch redhat-excluding-Reverts-from-changelog-too.patch redhat-updating-lastcommit-for-2-6-32-4.patch redhat-updating-lastcommit-for-2-6-32-5.patch redhat-check-if-patchutils-is-installed-before-creating-patches.patch redhat-do-a-basic-sanity-check-to-verify-the-modules-are-being-signed.patch redhat-updating-lastcommit-for-2-6-32-6.patch redhat-updating-lastcommit-for-2-6-32-7.patch redhat-updating-lastcommit-for-2-6-32-8.patch redhat-updating-lastcommit-for-2-6-32-9.patch redhat-updating-lastcommit-for-2-6-32-10.patch redhat-updating-lastcommit-for-2-6-32-11.patch redhat-updating-lastcommit-for-2-6-32-12.patch redhat-updating-lastcommit-for-2-6-32-13.patch redhat-updating-lastcommit-for-2-6-32-14.patch redhat-updating-lastcommit-for-2-6-32-15.patch redhat-updating-lastcommit-for-2-6-32-16.patch redhat-updating-lastcommit-for-2-6-32-17.patch redhat-updating-lastcommit-for-2-6-32-18.patch linux-kernel-test.patch"
+		for file in $(echo "${rh_crap_patch}" | tr ' ' '\n'); do
+			[ -e "${CWD}/${file}" ] && rm -rf "${CWD}/${file}" # >/dev/null 2>&1
+		done;
+
+		rm -rf "${CTD}" || die "${RED}rm -rf ${CTD} failed${NORMAL}"
+
+		cat kernel.spec | sed '1,/make -f %{SOURCE20} VERSION=%{version} configs/d; /ApplyOptionalPatch linux-kernel-test.patch/,$d' | sed 's/ApplyPatch //' > "${CWD}"/patch_list
+		for file in $(echo "${rh_crap_patch}" | tr ' ' '\n'); do
+			sed -i "/$file/d" "${CWD}"/patch_list;
+		done
+	elif [[ ${KMV} = "3.10" ]]; then
+		test -d "${WORKDIR}/linux-${KV_FULL}" >/dev/null 2>&1 || mkdir -p "${WORKDIR}/linux-${KV_FULL}"
+
+		rpm_unpack "${RH_NAME}.src.rpm" || die
+
+		if [ -e "linux-3.10.0-${RH_VER/KMV/$KMV}.el7.tar.xz" ]; then
+			tar -xpf "linux-3.10.0-${RH_VER/KMV/$KMV}.el7.tar.xz" || die
+			move "linux-3.10.0-${RH_VER/KMV/$KMV}.el7" "${S}"
+		fi
+
+		rm -rf "${CTD}" || die "${RED}rm -rf ${CTD} failed${NORMAL}"
 	fi
-
-	# Delete crap patches from rh patchset
-	local rh_crap_patch="linux-2.6.32.tar.bz2 redhat-Include-FIPS-required-checksum-of-the-kernel-image.patch redhat-Silence-tagging-messages-by-rh-release.patch redhat-Disabling-debug-options-for-beta.patch redhat-updating-lastcommit-for-2-6-31-50.patch redhat-fix-BZ-and-CVE-info-printing-on-changelog-when-HIDE_REDHAT-is-enabled.patch redhat-updating-lastcommit-for-2-6-31-51.patch redhat-Fix-version-passed-to-update_changelog-sh.patch redhat-updating-lastcommit-for-2-6-32-0-52.patch redhat-fix-STAMP-version-on-rh-release-commit-phase.patch redhat-enable-debug-builds-also-on-s390x-and-ppc64.patch redhat-updating-lastcommit-for-2-6-32-0-53.patch redhat-fixing-wrong-bug-number-536759-536769.patch redhat-adding-top-makefile-to-enable-rh-targets.patch redhat-updating-lastcommit-for-2-6-32-0-54.patch redhat-updating-lastcommit-for-2-6-32-1.patch redhat-update-build-targets-in-Makefile.patch redhat-include-missing-System-map-file-for-debug-only-builds.patch redhat-updating-lastcommit-for-2-6-32-2.patch redhat-force-to-run-rh-key-target-when-compiling-the-kernel-locally-without-RPM.patch redhat-fixing-lastcommit-contents-for-2-6-32-2-el6.patch redhat-updating-lastcommit-for-2-6-32-3.patch redhat-excluding-Reverts-from-changelog-too.patch redhat-updating-lastcommit-for-2-6-32-4.patch redhat-updating-lastcommit-for-2-6-32-5.patch redhat-check-if-patchutils-is-installed-before-creating-patches.patch redhat-do-a-basic-sanity-check-to-verify-the-modules-are-being-signed.patch redhat-updating-lastcommit-for-2-6-32-6.patch redhat-updating-lastcommit-for-2-6-32-7.patch redhat-updating-lastcommit-for-2-6-32-8.patch redhat-updating-lastcommit-for-2-6-32-9.patch redhat-updating-lastcommit-for-2-6-32-10.patch redhat-updating-lastcommit-for-2-6-32-11.patch redhat-updating-lastcommit-for-2-6-32-12.patch redhat-updating-lastcommit-for-2-6-32-13.patch redhat-updating-lastcommit-for-2-6-32-14.patch redhat-updating-lastcommit-for-2-6-32-15.patch redhat-updating-lastcommit-for-2-6-32-16.patch redhat-updating-lastcommit-for-2-6-32-17.patch redhat-updating-lastcommit-for-2-6-32-18.patch linux-kernel-test.patch"
-	for file in $(echo "${rh_crap_patch}" | tr ' ' '\n'); do
-		[ -e "${CWD}/${file}" ] && rm -rf "${CWD}/${file}" # >/dev/null 2>&1
-	done;
-
-	rm -rf "${CTD}" || die "${RED}rm -rf ${CTD} failed${NORMAL}"
-
-	cat kernel.spec | sed '1,/make -f %{SOURCE20} VERSION=%{version} configs/d; /ApplyOptionalPatch linux-kernel-test.patch/,$d' | sed 's/ApplyPatch //' > "${CWD}"/patch_list
-	for file in $(echo "${rh_crap_patch}" | tr ' ' '\n'); do
-		sed -i "/$file/d" "${CWD}"/patch_list;
-	done
 }
 
 # @FUNCTION: src_prepare
@@ -105,10 +125,13 @@ geek-rh_src_unpack() {
 geek-rh_src_prepare() {
 	debug-print-function ${FUNCNAME} "$@"
 
-	# now 2.6.32
-	ApplyPatch "${T}/rh/patch_list" "${RH_INF}"
-	# now 2.6.32.18
-	move "${T}/rh" "${WORKDIR}/linux-${KV_FULL}-patches/rh"
+	if [[ ${KMV} = "2.6" ]]; then
+		# now 2.6.32
+		ApplyPatch "${T}/rh/patch_list" "${RH_INF}"
+		# now 2.6.32.18
+		move "${T}/rh" "${WORKDIR}/linux-${KV_FULL}-patches/rh"
+#	elif [[ ${KMV} = "3.10" ]]; then
+	fi
 
 	# Comment out EXTRAVERSION added by rh patch:
 	sed -i -e "s:^\(EXTRAVERSION =\).*:\1 ${EXTRAVERSION}:" "Makefile" || die
